Convert Axion landing into the agency-landing template, rebrand to Hirael
Promotes the standalone /axion page into a first-class registry template
(category: templates) so it installs via the shadcn CLI and lands under
components/templates/agency-landing/ in a consumer repo.

- Restructured into registry/hirael/templates/agency-landing/ (root + hero +
  about + case-studies + primitives + shader-background), mirroring the
  creative-studio split. Declared in registry-meta.ts, added a preview loader
  in registry-demos.tsx, and removed the app/axion route. The WebGL shader
  stays behind dynamic(ssr: false).
- Rebranded to Hirael: inlined the arch-and-star brand mark as the nav logo,
  a "Hirael" wordmark, a "Hirael Studio" eyebrow, and an "Introducing Hirael"
  badge.
- Our projects: both cards now share one 4/3 aspect ratio, so the media is
  equal height across the two-up grid.
- About: replaced the overflowing whitespace-nowrap center column with a
  bottom-aligned baseline grid (small image, copy, large image) that no longer
  overlaps.

Regenerated registry.json / registry-props.json and updated docs. check:registry,
typecheck, lint, and the full static export all pass.
