+++ registry/hirael/registry-meta.ts
@@ -1357,6 +1357,31 @@ export const REGISTRY: RegistryEntryMeta[] = [
     registryDependencies: [],
     dependencies: ["framer-motion", "lucide-react"],
   },
+  {
+    name: "agency-landing",
+    title: "Agency Landing",
+    description:
+      "Bright, shader-lit agency landing page: a full-viewport hero with an animated WebGL backdrop, pill navigation and a live clock, an editorial about section, and a featured-work grid of autoplaying video cards. Built on the shaders package, with a self-contained light palette.",
+    category: "templates",
+    sourceFiles: [
+      "registry/hirael/templates/agency-landing/agency-landing.tsx",
+      "registry/hirael/templates/agency-landing/hero.tsx",
+      "registry/hirael/templates/agency-landing/about.tsx",
+      "registry/hirael/templates/agency-landing/case-studies.tsx",
+      "registry/hirael/templates/agency-landing/primitives.tsx",
+      "registry/hirael/templates/agency-landing/shader-background.tsx",
+    ],
+    installTargets: [
+      "components/templates/agency-landing/agency-landing.tsx",
+      "components/templates/agency-landing/hero.tsx",
+      "components/templates/agency-landing/about.tsx",
+      "components/templates/agency-landing/case-studies.tsx",
+      "components/templates/agency-landing/primitives.tsx",
+      "components/templates/agency-landing/shader-background.tsx",
+    ],
+    registryDependencies: [],
+    dependencies: ["shaders", "lucide-react"],
+  },
 ]
 
 /**
