+++ registry/hirael/templates/agency-landing/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./agency-landing"
