+++ registry/hirael/templates/agency-landing/primitives.tsx
@@ -1,0 +1,118 @@
+import { type ReactNode } from "react"
+import { ArrowRight } from "lucide-react"
+
+import { cn } from "@/lib/utils"
+
+// House easing for the text-roll / arrow-rotate hover motion.
+export const EASE = "ease-[cubic-bezier(0.25,0.1,0.25,1)]"
+
+/** Hirael arch-and-star brand mark. Strokes track the current text color. */
+export function HiraelMark({ className }: { className?: string }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 80 100"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2.4"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      aria-hidden="true"
+      className={className}
+    >
+      <path d="M16 78 V40 a24 24 0 0 1 48 0 V78" />
+      <path d="M40 44 L43.2 52 L51 55 L43.2 58 L40 66 L36.8 58 L29 55 L36.8 52 Z" />
+      <path d="M22 86 H58" opacity="0.7" />
+      <path d="M28 92 H52" opacity="0.45" />
+      <path d="M34 96 H46" opacity="0.25" />
+    </svg>
+  )
+}
+
+/**
+ * Hover text-roll: two stacked copies inside a 20px window that slide up by
+ * exactly one line on group-hover.
+ */
+export function RollText({ children }: { children: ReactNode }) {
+  return (
+    <span className="block h-[20px] overflow-hidden">
+      <span
+        className={cn(
+          "flex flex-col transition-transform duration-500 group-hover:-translate-y-1/2",
+          EASE
+        )}
+      >
+        <span className="block h-[20px] leading-[20px]">{children}</span>
+        <span className="block h-[20px] leading-[20px]" aria-hidden="true">
+          {children}
+        </span>
+      </span>
+    </span>
+  )
+}
+
+/** Orange pill CTA with the shared text-roll + arrow-rotate hover. */
+export function OrangeButton({
+  label,
+  className,
+}: {
+  label: string
+  className?: string
+}) {
+  return (
+    <button
+      type="button"
+      className={cn(
+        "group inline-flex items-center gap-2 rounded-full bg-[#F26522] py-2 pl-5 pr-2 text-[13px] font-medium text-white transition-colors hover:bg-[#e05a1a] sm:pl-6 sm:text-[14px]",
+        className
+      )}
+    >
+      <RollText>{label}</RollText>
+      <span className="flex h-7 w-7 items-center justify-center rounded-full bg-white sm:h-8 sm:w-8">
+        <ArrowRight
+          size={16}
+          className={cn(
+            "text-[#F26522] transition-transform duration-500 group-hover:-rotate-45",
+            EASE
+          )}
+        />
+      </span>
+    </button>
+  )
+}
+
+/** lucide "link" icon, inlined as two arc paths. */
+export function LinkIcon({ className }: { className?: string }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="14"
+      height="14"
+      viewBox="0 0 24 24"
+      fill="none"
+      stroke="currentColor"
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      className={className}
+      aria-hidden="true"
+    >
+      <path d="M10 13a5 5 0 0 0 7.54.54l3-3a5 5 0 0 0-7.07-7.07l-1.72 1.71" />
+      <path d="M14 11a5 5 0 0 0-7.54-.54l-3 3a5 5 0 0 0 7.07 7.07l1.71-1.71" />
+    </svg>
+  )
+}
+
+/** Certified-partner starburst mark. */
+export function PartnerIcon({ className }: { className?: string }) {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 100 100"
+      className={className}
+      aria-hidden="true"
+    >
+      <path d="m19.6 66.5 19.7-11 .3-1-.3-.5h-1l-3.3-.2-11.2-.3L14 53l-9.5-.5-2.4-.5L0 49l.2-1.5 2-1.3 2.9.2 6.3.5 9.5.6 6.9.4L38 49.1h1.6l.2-.7-.5-.4-.4-.4L29 41l-10.6-7-5.6-4.1-3-2-1.5-2-.6-4.2 2.7-3 3.7.3.9.2 3.7 2.9 8 6.1L37 36l1.5 1.2.6-.4.1-.3-.7-1.1L33 25l-6-10.4-2.7-4.3-.7-2.6c-.3-1-.4-2-.4-3l3-4.2L28 0l4.2.6L33.8 2l2.6 6 4.1 9.3L47 29.9l2 3.8 1 3.4.3 1h.7v-.5l.5-7.2 1-8.7 1-11.2.3-3.2 1.6-3.8 3-2L61 2.6l2 2.9-.3 1.8-1.1 7.7L59 27.1l-1.5 8.2h.9l1-1.1 4.1-5.4 6.9-8.6 3-3.5L77 13l2.3-1.8h4.3l3.1 4.7-1.4 4.9-4.4 5.6-3.7 4.7-5.3 7.1-3.2 5.7.3.4h.7l12-2.6 6.4-1.1 7.6-1.3 3.5 1.6.4 1.6-1.4 3.4-8.2 2-9.6 2-14.3 3.3-.2.1.2.3 6.4.6 2.8.2h6.8l12.6 1 3.3 2 1.9 2.7-.3 2-5.1 2.6-6.8-1.6-16-3.8-5.4-1.3h-.8v.4l4.6 4.5 8.3 7.5L89 80.1l.5 2.4-1.3 2-1.4-.2-9.2-7-3.6-3-8-6.8h-.5v.7l1.8 2.7 9.8 14.7.5 4.5-.7 1.4-2.6 1-2.7-.6-5.8-8-6-9-4.7-8.2-.5.4-2.9 30.2-1.3 1.5-3 1.2-2.5-2-1.4-3 1.4-6.2 1.6-8 1.3-6.4 1.2-7.9.7-2.6v-.2H49L43 72l-9 12.3-7.2 7.6-1.7.7-3-1.5.3-2.8L24 86l10-12.8 6-7.9 4-4.6-.1-.5h-.3L17.2 77.4l-4.7.6-2-2 .2-3 1-1 8-5.5Z" />
+    </svg>
+  )
+}
