+++ registry/hirael/templates/agency-landing/shader-background.tsx
@@ -3,10 +3,10 @@
 import { ChromaFlow, FilmGrain, FlutedGlass, Shader, Swirl } from "shaders/react"
 
 /**
- * The hero shader stack. Composited bottom-to-top by the root `<Shader>`
- * canvas: a soft swirl, the orange chroma flow, a fluted-glass refraction
- * pass, and a light film grain on top. Loaded client-only (it needs WebGL),
- * so it never runs during the static export.
+ * The hero shader stack, composited bottom-to-top by the root Shader canvas:
+ * a soft swirl, an orange chroma flow, a fluted-glass refraction pass, and a
+ * light film grain. WebGL only, so it loads client-side via the dynamic
+ * import in hero.tsx and never runs during server render.
  */
 export function ShaderBackground() {
   return (
